trunk: Fix SIL builds
git-svn-id: https://gentlenav.googlecode.com/svn/trunk@2533 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/navigate.c
@@ -87,7 +87,7 @@ void set_goal( struct relative3D fromPoint , struct relative3D toPoint )
 {
 	struct relative2D courseLeg ;
 
-	int courseDirection[2] ;
+	int16_t courseDirection[2] ;
 	
 	goal.x = toPoint.x ;
 	goal.y = toPoint.y ;
